BACKEND/requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==4.1.13
django-cors-headers==4.3.1
django-tinymce==4.0.0
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
et-xmlfile==1.1.0
filelock==3.13.4
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
platformdirs==4.2.0
pymongo==4.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.2.4
typing_extensions==4.11.0
tzdata==2024.1
virtualenv==20.25.1
